Reduce quality metrics collector coupling
Split the monitoring quality metrics collector into focused namespace collectors while keeping the existing public facade stable. Add a size-budget regression test for issue 597 and refresh graphify output after the code split.

Constraint: Issue #597 requires sub-500-line units and no quality batch regression.

Rejected: Keeping all collection and aggregation logic in quality-metrics-collector.ts | it preserves the tech-debt hotspot and fails the size budget.

Confidence: high

Scope-risk: moderate

Directive: Keep QualityMetricsCollector as the public orchestration facade; add namespace behavior in focused collectors.

Tested: npm test -- packages/memento-core/src/domains/monitoring/services/quality-assurance/quality-metrics-collector.spec.ts packages/memento-core/src/infrastructure/scheduler/jobs/__tests__/quality-measurement-batch-job.spec.ts; npm run type-check; npm run lint; graphify rebuild; git diff --check

Not-tested: Full repository npm test suite
